refactor: replace rawDataController with metricsController for log processing and streamline data routes
